// src/common/codegen/Voter.types.ts
export type Decimal = string;
export type Uint128 = string;
export interface AstroStakingRewardConfig {
  treasury: number;
  users: number;
}
export interface WeightAllocationItem {
  lp_token: string;
  weight: Decimal;
}
// ...
export interface BribesAllocationItem {
  lp_token: string;
  rewards: RewardsItem[];
}
export interface RewardsItem {
  amount: Uint128;
  symbol: string;
}